requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy
pandas
scikit-learn>=1.2
matplotlib
seaborn
pillow
tqdm
jupyter
notebook
scikit-image
faiss-cpu
joblib
mediapipe
requests